+++ aws/fedora-coreos/kubernetes/workers/ami.tf
@@ -23,6 +23,8 @@ data "aws_ami" "fedora-coreos" {
 # WARNING: These AMIs will be removed when Fedora CoreOS publishes arm64 AMIs
 # and may be removed for any reason before then as well. Do not use.
 data "aws_ami" "fedora-coreos-arm" {
+  count = var.arch == "arm64" ? 1 : 0
+
   most_recent = true
   owners      = ["099663496933"]